requirements.txt
factory_boy==2.12.0
Django==2.2.6
django-filter==2.2.0
django-fsm==2.6.1
djangorestframework==3.10.3
psycopg2-binary==2.8.4
pytz==2019.3
sqlparse==0.3.0
coverage==4.5.4
parameterized==0.7.0
